requirements.txt
asgiref==3.4.1
Django==3.2.8
django-split-settings==1.1.0
pytz==2021.3
sqlparse==0.4.2
